[Tizen] Add codes for Dali Windows Backend
This reverts commit cb92c9b7cb2d3ff05a688a4bd4af87c07f87084e.

+++ dali-toolkit/internal/layouting/linear-layout-impl.cpp
@@ -298,7 +298,7 @@ void LinearLayout::LayoutHorizontal( LayoutLength left, LayoutLength top, Layout
   auto height = bottom - top;
 
   // Space available for child
-  auto childSpace = height - padding.top - padding.bottom;
+  auto childSpace = height - (int)padding.top - (int)padding.bottom;
 
   auto count = GetChildCount();
 
@@ -360,24 +360,25 @@ void LinearLayout::LayoutHorizontal( LayoutLength left, LayoutLength top, Layout
       {
         case Dali::Toolkit::LinearLayout::Alignment::TOP:
         {
-          childTop = LayoutLength( padding.top ) + childMargin.top;
+          childTop = LayoutLength( padding.top ) + (int)childMargin.top;
           break;
         }
         case Dali::Toolkit::LinearLayout::Alignment::BOTTOM:
         {
-          childTop = height - padding.bottom - childHeight - childMargin.bottom;
+          childTop = height - (int)padding.bottom - childHeight - (int)childMargin.bottom;
           break;
         }
         case Dali::Toolkit::LinearLayout::Alignment::CENTER_VERTICAL:
         default:
         {
-          childTop = LayoutLength( padding.top ) + ( ( childSpace - childHeight ) / 2 ) + childMargin.top - childMargin.bottom;
+          childTop = LayoutLength( padding.top ) + (int)( ( childSpace - childHeight ) / 2 ) + (int)childMargin.top - (int)childMargin.bottom;
           break;
         }
       }
+
       childLeft += childMargin.start;
       childLayout->Layout( childLeft, childTop, childLeft + childWidth, childTop + childHeight );
-      childLeft += childWidth + childMargin.end + mCellPadding.width;
+      childLeft += childWidth + (int)childMargin.end + mCellPadding.width;
     }
   }
 }
@@ -508,7 +509,7 @@ void LinearLayout::LayoutVertical( LayoutLength left, LayoutLength top, LayoutLe
   auto width = right - left;
 
   // Space available for child
-  auto childSpace = width - padding.start - padding.end;
+  auto childSpace = width - (int)padding.start - (int)padding.end;
   auto count = GetChildCount();
 
   switch ( mAlignment & VERTICAL_ALIGNMENT_MASK )
@@ -564,7 +565,7 @@ void LinearLayout::LayoutVertical( LayoutLength left, LayoutLength top, LayoutLe
         }
       }
       childLayout->Layout( childLeft, childTop, childLeft + childWidth, childTop + childHeight );
-      childTop += childHeight + childMargin.bottom + mCellPadding.height;
+      childTop += childHeight + (int)childMargin.bottom + mCellPadding.height;
     }
   }
 }
